+++ canvas/index.js
@@ -1,0 +1,71 @@
+export { computeAutoLayout, computeTreeLayout } from './AutoLayout.js';
+export { SubgraphManager } from './SubgraphManager.js';
+export { SubgraphRouter } from './SubgraphRouter.js';
+export { LODManager } from './LODManager.js';
+export { PinExpansion } from './PinExpansion.js';
+export { ForceLayout } from './ForceLayout.js';
+export { FlowSimulator } from './FlowSimulator.js';
+export { routePcbTrace } from './PcbRouter.js';
+export { analyzePcbRoute, analyzePcbRouteSet, summarizePcbRouteQuality } from './PcbRouteDiagnostics.js';
+export {
+  createCanvasGraphStore,
+  normalizeCanvasGraphModel,
+} from './graph-model.js';
+export {
+  computeInitialGraphPositions,
+  createForceLayoutPayload,
+  findForceNodeGroup,
+  getDrillableFiles,
+  getForceLayoutOptions,
+  getGraphCacheKey,
+  getOrBuildGraph,
+} from './graph-layout.js';
+export {
+  GRAPH_DIRECTORY_FRAME_COLORS,
+  GRAPH_PATH_STYLES,
+  addGraphDirectoryFrames,
+  buildFlatPathHash,
+  getFileSelectionNodeId,
+  getFlatFocusRestoreKey,
+  getGraphHashNavigationState,
+  getGraphPathStyleDisplay,
+  getNextGraphPathStyle,
+  renderGraphPathStyleButton,
+  renderGraphViewModeButton,
+  resolveFlatHashChange,
+  resolveInitialGraphViewMode,
+  selectGraphLabelMode,
+  setGraphLayerVisible,
+  shouldClearFocusOnSelection,
+  shouldFitForceLayoutInitialTick,
+  shouldRestoreFlatFocus,
+  toggleGraphLayerButtonState,
+  resolveGraphNodeClick,
+  resolveToolbarAction,
+  renderClusterPanel,
+  renderGraphStats,
+} from './graph-explorer.js';
+export {
+  buildCanvasGraphModelFromSkeleton,
+  buildGraphModelFromSkeleton,
+} from './project-graph-model.js';
+export {
+  buildFileGraph,
+  buildStructuredGraph,
+} from './project-graph-builder.js';
+export {
+  HTML_IN_CANVAS_APIS,
+  HTML_IN_CANVAS_RENDERER,
+  HTML_IN_CANVAS_RENDERER_NAME,
+  captureHtmlElementImage,
+  closeHtmlElementImage,
+  copyHtmlElementToWebGPUTexture,
+  createHtmlInCanvasAdapter,
+  drawHtmlElement2d,
+  getHtmlInCanvasChangedElements,
+  getHtmlElementCanvasTransform,
+  getHtmlInCanvasSupport,
+  requestHtmlInCanvasPaint,
+  setupHtmlInCanvas,
+  uploadHtmlElementToWebGLTexture,
+} from './html-in-canvas.js';